+++ code/algorithm-demo/src/com/noob/algorithm/leetcode/q002/Main.java
@@ -1,0 +1,15 @@
+package com.noob.algorithm.leetcode.q002;
+
+import com.noob.algorithm.base.dataStructure.tree.ListNode;
+
+public class Main {
+
+    public static void main(String[] args) {
+
+        // 构建测试用例
+        ListNode node1 = new ListNode(-1);
+
+        // Solution1 solution1 = new Solution1();
+
+    }
+}
